+++ src/main/java/com/packsendme/lib/simulation/way/airway/response/dto/SimulationContainerAirlineResponse_Dto.java
@@ -1,4 +1,4 @@
-package com.packsendme.lib.simulation.to.airway.response.dto;
+package com.packsendme.lib.simulation.way.airway.response.dto;
 
 import java.io.Serializable;
