+++ src/main/java/com/packsendme/lib/simulation/way/roadway/response/dto/SimulationBicycleResponse_Dto.java
@@ -1,4 +1,4 @@
-package com.packsendme.lib.simulation.to.roadway.response.dto;
+package com.packsendme.lib.simulation.way.roadway.response.dto;
 
 import java.io.Serializable;
 
@@ -9,11 +9,12 @@ import lombok.Setter;
 @Getter 
 @Setter 
 @NoArgsConstructor
-public class SimulationWalkResponse_Dto implements Serializable  {
+public class SimulationBicycleResponse_Dto implements Serializable  {
 	
 	/**
 	 * 
 	 */
 	public static final long serialVersionUID = 1L;
-	public Double cost_walk;
+	public Double cost_bicycle;
+	
 }
